# swagger.yaml
openapi: 3.0.0
info:
  title: OttoHUB Atm API
  version: 1.0.0
  description: API for managing ATM operations
  contact:
    name: Martin Negin, Samu Lyhty, Veikka Koskinen, Topias Perälä
    email: your.email@example.com
servers:
  - url: https://atm.dev.zroot.it
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /currency:
    get:
      summary: Get a list of currencies
      description: Retrieve a list of all available currencies
      responses:
        '200':
          description: A list of currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The currency ID
                    currency_type:
                      type: string
                      description: The type of currency (e.g., USD, EUR)
                    currency_value:
                      type: number
                      description: The value of the currency
                    currency_symbol:
                      type: string
                      description: The symbol of the currency (e.g., $, €)
        '500':
          description: Internal server error
  /currency/now:
    get:
      summary: Update currency rates immediately
      description: Fetch and update the latest currency rates immediately
      responses:
        '200':
          description: Currency rates updated successfully
        '500':
          description: Internal server error
...